finish refactor for pricing-table block. relates to #504.
