+++ src/blocks/pricing-table/index.js
@@ -1,86 +1,47 @@
 /**
- * External dependencies
- */
-import classnames from 'classnames';
-
-/**
  * Internal dependencies
  */
 import './styles/editor.scss';
 import './styles/style.scss';
-import edit from './components/edit';
+import edit from './edit';
 import icons from './../../utils/icons';
+import save from './save';
+import transforms from './transforms';
 import metadata from './block.json';
 
 /**
  * WordPress dependencies
  */
 const { __ } = wp.i18n;
-const { createBlock } = wp.blocks;
-const { InnerBlocks } = wp.blockEditor;
 
 /**
  * Block constants
  */
-const { name } = metadata;
+const { attributes, name } = metadata;
+
+const title = __( 'Pricing Table' );
 
 const icon = icons.pricing;
 
 const settings = {
-	title: __( 'Pricing Table' ),
+	title,
+
 	description: __( 'Add pricing tables.' ),
+
 	keywords: [ __( 'landing' ), __( 'comparison' ), __( 'coblocks' ) ],
-	attributes: metadata.attributes,
+
+	attributes,
+
 	supports: {
 		align: [ 'wide', 'full' ],
 		html: false,
 	},
-	transforms: {
-		from: [
-			{
-				type: 'raw',
-				selector: 'div.wp-block-coblocks-pricing-table',
-				schema: {
-					div: {
-						classes: [ 'wp-block-coblocks-pricing-table' ],
-					},
-				},
-			},
-			{
-				type: 'prefix',
-				prefix: ':pricing',
-				transform: function( content ) {
-					return createBlock( `coblocks/${ name }`, {
-						content,
-					} );
-				},
-			},
-		],
-	},
+
+	transforms,
+
 	edit,
-	save( { attributes } ) {
-		const {
-			contentAlign,
-			count,
-		} = attributes;
 
-		const classes = classnames(
-			`has-${ count }-columns`,
-			`has-${ contentAlign }-content`,
-		);
-
-		return (
-
-			<div
-				className={ classes }
-				style={ { textAlign: contentAlign ? contentAlign : null } }
-			>
-				<div className="wp-block-coblocks-pricing-table__inner">
-					<InnerBlocks.Content />
-				</div>
-			</div>
-		);
-	},
+	save,
 };
 
 export { name, icon, settings };
