+++ src/blocks/pricing-table/inspector.js
