+++ src/blocks/pricing-table/pricing-table-item/inspector.js
