+++ src/blocks/pricing-table/pricing-table-item/save.js
@@ -1,0 +1,79 @@
+/**
+ * External dependencies
+ */
+import classnames from 'classnames';
+
+/**
+ * WordPress dependencies
+ */
+const { RichText, getColorClassName, InnerBlocks } = wp.blockEditor;
+
+const save = ( { attributes } ) => {
+	const {
+		amount,
+		currency,
+		customBackgroundColor,
+		customTextColor,
+		features,
+		backgroundColor,
+		textColor,
+		title,
+	} = attributes;
+
+	const backgroundClass = getColorClassName( 'background-color', backgroundColor );
+	const textColorClass = getColorClassName( 'color', textColor );
+
+	const classes = classnames( {
+		'has-background': backgroundColor || customBackgroundColor,
+		[ backgroundClass ]: backgroundClass,
+		'has-text-color': textColor || customTextColor,
+		[ textColorClass ]: textColorClass,
+	}
+	);
+
+	const styles = {
+		backgroundColor: backgroundClass ? undefined : customBackgroundColor,
+		color: textColorClass ? undefined : customTextColor,
+	};
+
+	return (
+		<div
+			className={ classes }
+			style={ styles }
+		>
+			{ ! RichText.isEmpty( title ) && (
+				<RichText.Content
+					tagName="span"
+					className="wp-block-coblocks-pricing-table-item__title"
+					value={ title }
+				/>
+			) }
+			{ ! RichText.isEmpty( amount ) && (
+				<div className={ 'wp-block-coblocks-pricing-table-item__price-wrapper' }>
+					{ ! RichText.isEmpty( currency ) && (
+						<RichText.Content
+							tagName="span"
+							className="wp-block-coblocks-pricing-table-item__currency"
+							value={ currency }
+						/>
+					) }
+					<RichText.Content
+						tagName="span"
+						className="wp-block-coblocks-pricing-table-item__amount"
+						value={ amount }
+					/>
+				</div>
+			) }
+			{ ! RichText.isEmpty( features ) && (
+				<RichText.Content
+					tagName="ul"
+					className="wp-block-coblocks-pricing-table-item__features"
+					value={ features }
+				/>
+			) }
+			<InnerBlocks.Content />
+		</div>
+	);
+};
+
+export default save;
