+++ src/blocks/pricing-table/pricing-table-item/transforms.js
@@ -1,0 +1,15 @@
+const transforms = {
+	from: [
+		{
+			type: 'raw',
+			selector: 'div.wp-block-coblocks-pricing-table',
+			schema: {
+				div: {
+					classes: [ 'wp-block-coblocks-pricing-table' ],
+				},
+			},
+		},
+	],
+};
+
+export default transforms;
